requirements.txt
# Core ML
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0
xgboost==2.0.3
yfinance==0.2.36
prophet==1.1.5
matplotlib==3.8.2

# MLOps
dvc==3.42.0
dvc-s3==3.0.1
mlflow==2.10.2
dagshub==0.3.24
python-dotenv==1.0.0

# Deployment
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.0
requests==2.31.0
